Conf

Conf

JDP.Conf.confmerge

JDP.Conf.get_conf

JDP.Conf.set_conf

JDP.Conf.set_usr_path

JDP.Conf — Module.

Gives access to configuration files

source

JDP.Conf.confmerge — Method.

Like Base.merge, but recurses into Dictionaries

source

JDP.Conf.get_conf — Method.

get_conf(name::Symbol)::Dict

Get the configuration for name. If a temporary in-memory conf has been set with set_conf then it will return that. Otherwise it will return the contents of ../conf/name.toml merged with ~/.config/jdp/name.toml. The contents of the home directory config win in the event of a conflict.

source

JDP.Conf.set_conf — Method.

Used to override the contents of the configuration files for testing (for now)

source

JDP.Conf.set_usr_path — Method.

Override the user specific config path for unit testing

source