Repository

Repository

JDP.Repository.AbstractItem

JDP.Repository.fetch

JDP.Repository.refresh

JDP.Repository.AbstractItem — Type.

Some kind of item tracked by a tracker

source

JDP.Repository.fetch — Method.

Get one or more items of the given in the specified container

The exact behaviour depends on what is requested. If the data can not be retrieved from the local data cache then it may request it from a remote source.

source

JDP.Repository.refresh — Method.

Refresh the local cached data for the given bug references

source