Trackers

Bug References

JDP.Tracker.AbstractSession

JDP.Tracker.Api

JDP.Tracker.Instance

JDP.Tracker.Instance

JDP.Tracker.StaticSession

JDP.Tracker.TrackerRepo

JDP.Tracker.ensure_login!

JDP.Trackers.OpenQA.NativeSession

JDP.Trackers.OpenQA.Session

Tracker

JDP.Tracker — Module.

Trackers are external sources of information which track some kind of item.

For example Bugzilla and OpenQA are both considered trackers by JDP. Bugzilla tracks bugs and OpenQA tracks test results. GitWeb could also be considered a tracker which tracks git commits. Some services may track a number of different items.

Design note

Hopefully new trackers can eventually be declaratively defined in conf/trackers.toml. However this is difficult when most of them seem to use different authentication methods and different data formats. So we begin with tracker specific code (e.g. trackers/Bugzilla.jl) and then try to generialise them if feasible.

source

JDP.Tracker.Api — Type.

Information about a tracker's API

This is a generic interface for tracker features which are simple/standard enough to configure via conf/trackers.toml. Tracker specific features are handled by Tracker specific methods dispatched on the Session type parameter

source

JDP.Tracker.TrackerRepo — Type.

Tracker Repository

source

JDP.Tracker.AbstractSession — Type.

A connection to a tracker API

source

JDP.Tracker.Instance — Type.

Information about a Tracker's instance

source

JDP.Tracker.Instance — Method.

Create a minimal tracker instance for an unknown tracker

source

JDP.Tracker.StaticSession — Type.

Not really a session

source

JDP.Tracker.ensure_login! — Method.

Returns an active session

If the tracker already has an active session then return it, otherwise create one. The tracker specific modules should override this

source

Trackers.OpenQA

JDP.Trackers.OpenQA.NativeSession — Type.

Use native Julia HTTP library to access OpenQA

Unfortunately this doesn't work so well because:

A) JuliaWeb's current HTTP SSL implementation i.e. the MbedTLS wrapper B) OpenQA's wierd authentication which is difficult to replicate outside of Perl.

source

JDP.Trackers.OpenQA.Session — Type.

Makes requests to OpenQA using the official OpenQA client script

I really hate this, but we cache the data locally anyway due to the slowness of fetching from OpenQA, so the overhead of calling a Perl script can be ignored. Also see OpenQA::NativeSession's docs.

source

Trackers.Bugzilla